refactor: Remove duplication in members pages
