+++ src/collaboration/collaborationContext.js
@@ -42,12 +42,15 @@ export const CollaborationContextProvider = props => {
           'onMemberJoin',
           'onMemberRemove',
           'onMemberRoleChange',
+          'onMemberApprove',
           'MemberLeaveButton',
           'MemberRemoveButton',
           'MemberJoinButton',
           'MemberRequestJoinButton',
           'MemberRequestCancelButton',
           'updateOwner',
+          'acceptedRoles',
+          'allowedToManageMembers',
         ],
         props
       ),
